+++ src/serve.js
@@ -43,6 +43,7 @@ const MIME = {
   ".json": "application/json; charset=utf-8",
   ".png": "image/png",
   ".svg": "image/svg+xml",
+  ".bmp": "image/bmp",
 };
 
 async function loadManifest(dir) {